Properly calculate the summary.
It turns out calculating the summary is actually a little tricky.
You first need the fully rendered markdown to get rid of the markdown
syntax.
Then you need to try to extract the first paragraph from the resulting
html.

We take shortcut on the second part by just looking for the first,
closing paragraph tag and splitting on that.
This assumes that the resulting html always starts with a paragraph.