Switch to -C argument
Many programs have the ability to specify a directory from which the
entir program should be run (instead of the current directory). Let's
add that to kblog and get rid of all the other arguments. This
simplifies operation of the program by assuming the canoncial directory
structure found in the README.

Fixes #4